dev_requirements.txt
cx_Freeze
requests
selenium
setuptools
pyserial